Validation of planet object

+++ src/main/java/com/planetapi/business/PlanetBusinessImpl.java
@@ -1,8 +1,10 @@
 package com.planetapi.business;
 
 import com.planetapi.exception.PlanetAlreadyExistsException;
+import com.planetapi.exception.PlanetNotFoundException;
 import com.planetapi.model.Planet;
 import com.planetapi.repository.PlanetRepository;
+import com.planetapi.service.SwapiRequestService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,6 +20,8 @@ public class PlanetBusinessImpl implements PlanetBusiness {
     private PlanetRepository planetRepository;
     @Autowired
     private Validation validation;
+    @Autowired
+    private SwapiRequestService swapiRequestService;
 
     @Override
     public Planet addPlanet(Planet planet){
@@ -33,6 +37,8 @@ public class PlanetBusinessImpl implements PlanetBusiness {
 
         Planet insertedPlanet = planetRepository.insert(planet);
 
+        insertedPlanet.setApparitionCounter(swapiRequestService.getApparisons(insertedPlanet.getName()));
+
         return insertedPlanet;
     }
 
@@ -43,6 +49,8 @@ public class PlanetBusinessImpl implements PlanetBusiness {
 
         List<Planet> planets = planetRepository.findAll();
 
+        planets.forEach(planet -> planet.setApparitionCounter(swapiRequestService.getApparisons(planet.getName())));
+
         return planets;
     }
 
@@ -51,7 +59,15 @@ public class PlanetBusinessImpl implements PlanetBusiness {
 
         log.info("Getting planet with name {}", id);
 
+        validation.validateField("id", id);
+
         Optional<Planet> planetOptional = planetRepository.findById(id);
+
+        if (!planetOptional.isPresent())
+            throw new PlanetNotFoundException();
+
+        Planet planet = planetOptional.get();
+        planet.setApparitionCounter(swapiRequestService.getApparisons(planet.getName()));
 
         return planetOptional.get();
     }
@@ -61,15 +77,28 @@ public class PlanetBusinessImpl implements PlanetBusiness {
 
         log.info("Getting planet with name {}", name);
 
+        validation.validateField("name", name);
+
         Optional<Planet> planetOptional = planetRepository.findByName(name);
 
-        return planetOptional.get();
+        if (!planetOptional.isPresent())
+            throw new PlanetNotFoundException();
+
+        Planet planet = planetOptional.get();
+        planet.setApparitionCounter(swapiRequestService.getApparisons(planet.getName()));
+
+        return planet;
     }
 
     @Override
     public void deletePlanetById(String id){
 
         log.info("Deleting planet with id {}", id);
+
+        Optional<Planet> planetOptional = planetRepository.findById(id);
+
+        if (!planetOptional.isPresent())
+            throw new PlanetNotFoundException();
 
         planetRepository.deleteById(id);
     }
